LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-audioparam-interface/k-rate-delay-connections-expected.txt


PASS # AUDIT TASK RUNNER STARTED. 
PASS Executing "delayTime" 
PASS Audit report 
PASS > [delayTime] DelayNode delayTime k-rate input 
PASS   refNode = new DelayNode(context) did not throw an exception. 
PASS   refNode.delayTime.automationRate = 'k-rate' did not throw an exception. 
PASS   refNode.delayTime.setValueAtTime(0, 0) did not throw an exception. 
PASS   refNode.delayTime.linearRampToValueAtTime(1.125, 1) did not throw an exception. 
PASS   testNode = new DelayNode(context) did not throw an exception. 
PASS   testNode.delayTime.automationRate = 'k-rate' did not throw an exception. 
PASS   testMod = new ConstantSourceNode(context) did not throw an exception. 
PASS   testMod.offset.setValueAtTime(0, 0) did not throw an exception. 
PASS   testMod.offset.linearRampToValueAtTime(1.125, 1) did not throw an exception. 
PASS   testMod.connect(testNode.delayTime) did not throw an exception. 
PASS   Expected k-rate delayTime AudioParam with input is not constantly 0 (contains 6159 different values). 
PASS   Actual k-rate delayTime AudioParam with input is not constantly 0 (contains 6159 different values). 
FAIL X output[128, 143]: Expected 0 for all values but found 16 unexpected values: 
	Index	Actual
	[0]	-0.051677267998456955
	[1]	-0.0074212560430169106
	[2]	0.03865136206150055
	[3]	0.08652953058481216
	...and 12 more errors. assert_true: expected true got false
PASS   output[144, 255] is not constantly 0 (contains 112 different values). 
PASS   k-rate delayTime AudioParam with input equals [0,0.3387378752231598,0.6374238729476929,0.8607419729232788,0.9822871685028076,0.9876880645751953,0.8763062953948975,0.6613117456436157,0.3681243062019348,0.03141050040721893,-0.3090170621871948,-0.6129071116447449,-0.8443278074264526,-0.9759165644645691,-0.9921145439147949,-0.8910062313079834...] with an element-wise tolerance of {"absoluteThreshold":0,"relativeThreshold":0}. 
FAIL < [delayTime] 1 out of 15 assertions were failed. assert_true: expected true got false
FAIL # AUDIT TASK RUNNER FINISHED: 1 out of 1 tasks were failed. assert_true: expected true got false
